sec_fred_rag_app/requirements.txt
# ── Core ────────────────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.26.0

# ── SEC parsing ──────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=4.9.0
spacy>=3.7.0

# ── Embeddings + Vector DB ───────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.27.0

# ── FRED ─────────────────────────────────────────────────────
fredapi>=0.5.0

# ── Analytics ────────────────────────────────────────────────
vaderSentiment>=3.3.2
scikit-learn>=1.3.0
scipy>=1.11.0

# ── Visualization + App ──────────────────────────────────────
streamlit>=1.35.0
plotly>=5.18.0

dash>=4.0.0
dash-bootstrap-components>=2.0.0
dash-iconify>=0.1.2
